Break up FilesBuckets into smaller components
